Made Medium Template
Medium template almost done, just have to add some extra classes for the last few sems

+++ BE-Road2CS/src/main/java/com/road2cs/models/Semester.java
@@ -10,20 +10,20 @@ import java.util.Collection;
 @Getter
 @Setter
 public class Semester {
-    private int termNum; //which term it is
+    private double termNum; //which term it is
     private String termName; //fall/spring/summer/winter
     private Map<String, Course> courses;  //courseId to Course
     private int semUnits;
 
     //Constructor - initialize with empty HashMap
-    public Semester(int termNum, String termName) {
+    public Semester(double termNum, String termName) {
         this.termNum = termNum;
         this.termName = termName;
         this.courses = new HashMap<>();
     }
 
     //Constructor - initialize with existing courses
-    public Semester(int termNum, String termName, Map<String, Course> courses) {
+    public Semester(double termNum, String termName, Map<String, Course> courses) {
         this.termNum = termNum;
         this.termName = termName;
         this.courses = courses != null ? courses : new HashMap<>();
